Fix usage on non-termguicolors enabled vim
PR #93 introduced support for `termguicolors`.

Unfortunately the check for `termguicolors` is executed even if vim has
been compiled without `termguicolors` support (which is the case for
example for vim on Ubuntu 16.04 Xenial LTS), resulting in:

E113: Unknown option: termguicolors
E15: Invalid expression: &termguicolors

This patch fixes the shortcoming by checking for the existence of
`termguicolors` first.